Update RB config for new options
I am unsure of the difference between mode=beast and raw, but it seems like raw is used in the radarbox bare-metal install and it works in the current configuration, so I went with it.

Removed the now-unused mlat-client log directory

Explicitly disable MLAT if MLAT is turned off by the user

Start up the MLAT client via RB feeder because this seems to be the only configuration where MLAT actually ever gets fed to radarbox.